refactor: renamed everything to navigate easily
