+++ tests/autotests/test_task02.py
@@ -14,7 +14,7 @@ from grammars_constants import REGEXES
 
 # Fix import statements in try block to run tests
 try:
-    from project.finite_automaton_utils import regex_to_dfa, graph_to_nfa
+    from project.task2 import regex_to_dfa, graph_to_nfa
 except ImportError:
     pytestmark = pytest.mark.skip("Task 2 is not ready to test!")
