+++ tests/autotests/test_task03.py
@@ -12,12 +12,8 @@ from rpq_concrete_cases import CASES_RPQ, CaseRPQ
 
 # Fix import statements in try block to run tests
 try:
-    from project.finite_automaton_utils import regex_to_dfa
-    from project.AdjacencyMatrixFA import (
-        intersect_automata,
-        AdjacencyMatrixFA,
-        tensor_based_rpq,
-    )
+    from project.task2 import regex_to_dfa
+    from project.task3 import intersect_automata, AdjacencyMatrixFA, tensor_based_rpq
 except ImportError:
     pytestmark = pytest.mark.skip("Task 3 is not ready to test!")
