+++ tests/test_my_task2.py
@@ -4,8 +4,8 @@ import random
 import networkx as nx
 from pyformlang.finite_automaton import Symbol
 
-from project.finite_automaton_utils import regex_to_dfa, graph_to_nfa
-from project.graph_utils import create_two_cycles_graph_and_write_to_dot
+from project.task2 import regex_to_dfa, graph_to_nfa
+from project.task1 import create_two_cycles_graph_and_write_to_dot
 
 
 class TestRegexToDfa:
